% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefAdaptAge}
\alias{get_coefAdaptAge}
\title{coefAdaptAge}
\usage{
get_coefAdaptAge()
}
\value{
coefAdaptAge dataset
}
\description{
Coefficients AdaptAge adaptive causal clock. 998 CpGs described in Ying et
al. (2024).
}
\details{
\code{get_coefAdaptAge} retrieves this resource from ExperimentHub.
}
\examples{
get_coefAdaptAge()
}
