% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefAltumAge}
\alias{get_coefAltumAge}
\title{coefAltumAge}
\usage{
get_coefAltumAge()
}
\value{
path to the coefAltumAge file in the local cache
}
\description{
AltumAge neural network weights and scaler. Layer weights and input scaler
of the AltumAge neural network, from de Lima Camillo et al. (2022).
}
\details{
\code{get_coefAltumAge} retrieves this resource from ExperimentHub.
}
\examples{
get_coefAltumAge()
}
