% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefAltumAgeRef}
\alias{get_coefAltumAgeRef}
\title{coefAltumAgeRef}
\usage{
get_coefAltumAgeRef()
}
\value{
coefAltumAgeRef dataset
}
\description{
AltumAge input CpG order. The input CpG identifiers in the order the
AltumAge network expects, from de Lima Camillo et al. (2022).
}
\details{
\code{get_coefAltumAgeRef} retrieves this resource from ExperimentHub.
}
\examples{
get_coefAltumAgeRef()
}
