% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_coefBLUP}
\alias{get_coefBLUP}
\title{coefBLUP}
\usage{
get_coefBLUP()
}
\value{
coefBLUP dataset
}
\description{
\code{get_coefBLUP} returns the Best Linear Unbiased Prediction (BLUP) clock coefficients
}
\examples{
get_coefBLUP()
}
