% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_coefBohlin}
\alias{get_coefBohlin}
\title{coefBohlin}
\usage{
get_coefBohlin()
}
\value{
coefBohlin dataset
}
\description{
\code{get_coefBohlin} returns the Bohlin's clock coefficients
}
\examples{
cget_coefBohlinoefBohlin()
}
