% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefCausAge}
\alias{get_coefCausAge}
\title{coefCausAge}
\usage{
get_coefCausAge()
}
\value{
coefCausAge dataset
}
\description{
Coefficients CausAge causal clock. 581 CpGs described in Ying et al.
(2024).
}
\details{
\code{get_coefCausAge} retrieves this resource from ExperimentHub.
}
\examples{
get_coefCausAge()
}
