% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefDamAge}
\alias{get_coefDamAge}
\title{coefDamAge}
\usage{
get_coefDamAge()
}
\value{
coefDamAge dataset
}
\description{
Coefficients DamAge damaging causal clock. 1089 CpGs described in Ying et
al. (2024).
}
\details{
\code{get_coefDamAge} retrieves this resource from ExperimentHub.
}
\examples{
get_coefDamAge()
}
