% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefDunedinPACE}
\alias{get_coefDunedinPACE}
\title{coefDunedinPACE}
\usage{
get_coefDunedinPACE()
}
\value{
coefDunedinPACE dataset
}
\description{
Coefficients DunedinPACE pace-of-aging estimator. 173 CpGs described in
Belsky (2022).
}
\details{
\code{get_coefDunedinPACE} retrieves this resource from ExperimentHub.
}
\examples{
get_coefDunedinPACE()
}
