% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefDunedinPACEGS}
\alias{get_coefDunedinPACEGS}
\title{coefDunedinPACEGS}
\usage{
get_coefDunedinPACEGS()
}
\value{
coefDunedinPACEGS dataset
}
\description{
DunedinPACE reference means for probe normalization. Reference probe means
used to normalize the DunedinPACE input, from Belsky (2022).
}
\details{
\code{get_coefDunedinPACEGS} retrieves this resource from ExperimentHub.
}
\examples{
get_coefDunedinPACEGS()
}
