% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefEpiTOC1}
\alias{get_coefEpiTOC1}
\title{coefEpiTOC1}
\usage{
get_coefEpiTOC1()
}
\value{
coefEpiTOC1 dataset
}
\description{
CpG set for the epiTOC1 mitotic counter. 385 CpGs described in Yang et al.
(2016).
}
\details{
\code{get_coefEpiTOC1} retrieves this resource from ExperimentHub.
}
\examples{
get_coefEpiTOC1()
}
