% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefEpiTOC2}
\alias{get_coefEpiTOC2}
\title{coefEpiTOC2}
\usage{
get_coefEpiTOC2()
}
\value{
coefEpiTOC2 dataset
}
\description{
Parameters for the epiTOC2 mitotic counter. 163 CpGs described in
Teschendorff (2020).
}
\details{
\code{get_coefEpiTOC2} retrieves this resource from ExperimentHub.
}
\examples{
get_coefEpiTOC2()
}
