% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_coefHannum}
\alias{get_coefHannum}
\title{coefHannum}
\usage{
get_coefHannum()
}
\value{
coefHannum dataset
}
\description{
\code{get_coefHannum} returns the Hannum's clock coefficients
}
\examples{
get_coefHannum()
}
