% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_coefHorvath}
\alias{get_coefHorvath}
\title{coefHorvath}
\usage{
get_coefHorvath()
}
\value{
coefHorvath dataset
}
\description{
\code{get_coefHorvath} returns the Hobarth's clock coefficients
}
\examples{
get_coefHorvath()
}
