% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefHypoClock}
\alias{get_coefHypoClock}
\title{coefHypoClock}
\usage{
get_coefHypoClock()
}
\value{
coefHypoClock dataset
}
\description{
CpG set for the HypoClock solo-WCGW counter. 678 CpGs described in
Teschendorff (2020).
}
\details{
\code{get_coefHypoClock} retrieves this resource from ExperimentHub.
}
\examples{
get_coefHypoClock()
}
