% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_coefKnightGA}
\alias{get_coefKnightGA}
\title{coefKnightGA}
\usage{
get_coefKnightGA()
}
\value{
coefKnightGA dataset
}
\description{
\code{get_coefKnightGA} returns the Knight's clock coefficients
}
\examples{
get_coefKnightGA()
}
