% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_coefLeeGA}
\alias{get_coefLeeGA}
\title{coefLeeGA}
\usage{
get_coefLeeGA()
}
\value{
coefLeeGA dataset
}
\description{
\code{get_coefLeeGA} returns the Lee's Gestational Age clock coefficients
}
\examples{
get_coefLeeGA()
}
