% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_coefLevine}
\alias{get_coefLevine}
\title{coefLevine}
\usage{
get_coefLevine()
}
\value{
coefLevine dataset
}
\description{
\code{get_coefLevine} returns the Levine's clock coefficients
}
\examples{
get_coefLevine()
}
