% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefLin}
\alias{get_coefLin}
\title{coefLin}
\usage{
get_coefLin()
}
\value{
coefLin dataset
}
\description{
Coefficients Lin blood clock. 99 CpGs described in Lin et al. (2016).
}
\details{
\code{get_coefLin} retrieves this resource from ExperimentHub.
}
\examples{
get_coefLin()
}
