% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_coefMayneGA}
\alias{get_coefMayneGA}
\title{coefMayneGA}
\usage{
get_coefMayneGA()
}
\value{
coefMayneGA dataset
}
\description{
\code{get_coefMayneGA} returns the Mayne's clock coefficients
}
\examples{
get_coefMayneGA()
}
