% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefMcCartneyAlcohol}
\alias{get_coefMcCartneyAlcohol}
\title{coefMcCartneyAlcohol}
\usage{
get_coefMcCartneyAlcohol()
}
\value{
coefMcCartneyAlcohol dataset
}
\description{
Coefficients McCartney alcohol EpiScore. 450 CpGs described in McCartney et
al. (2018).
}
\details{
\code{get_coefMcCartneyAlcohol} retrieves this resource from ExperimentHub.
}
\examples{
get_coefMcCartneyAlcohol()
}
