% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefMcCartneyBMI}
\alias{get_coefMcCartneyBMI}
\title{coefMcCartneyBMI}
\usage{
get_coefMcCartneyBMI()
}
\value{
coefMcCartneyBMI dataset
}
\description{
Coefficients McCartney body mass index EpiScore. 1109 CpGs described in
McCartney et al. (2018).
}
\details{
\code{get_coefMcCartneyBMI} retrieves this resource from ExperimentHub.
}
\examples{
get_coefMcCartneyBMI()
}
