% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefMcCartneyBodyFat}
\alias{get_coefMcCartneyBodyFat}
\title{coefMcCartneyBodyFat}
\usage{
get_coefMcCartneyBodyFat()
}
\value{
coefMcCartneyBodyFat dataset
}
\description{
Coefficients McCartney body fat EpiScore. 968 CpGs described in McCartney
et al. (2018).
}
\details{
\code{get_coefMcCartneyBodyFat} retrieves this resource from ExperimentHub.
}
\examples{
get_coefMcCartneyBodyFat()
}
