% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefMcCartneyEducation}
\alias{get_coefMcCartneyEducation}
\title{coefMcCartneyEducation}
\usage{
get_coefMcCartneyEducation()
}
\value{
coefMcCartneyEducation dataset
}
\description{
Coefficients McCartney education EpiScore. 373 CpGs described in McCartney
et al. (2018).
}
\details{
\code{get_coefMcCartneyEducation} retrieves this resource from
ExperimentHub.
}
\examples{
get_coefMcCartneyEducation()
}
