% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefMcCartneyHDL}
\alias{get_coefMcCartneyHDL}
\title{coefMcCartneyHDL}
\usage{
get_coefMcCartneyHDL()
}
\value{
coefMcCartneyHDL dataset
}
\description{
Coefficients McCartney HDL cholesterol EpiScore. 737 CpGs described in
McCartney et al. (2018).
}
\details{
\code{get_coefMcCartneyHDL} retrieves this resource from ExperimentHub.
}
\examples{
get_coefMcCartneyHDL()
}
