% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefMcCartneyLDL}
\alias{get_coefMcCartneyLDL}
\title{coefMcCartneyLDL}
\usage{
get_coefMcCartneyLDL()
}
\value{
coefMcCartneyLDL dataset
}
\description{
Coefficients McCartney LDL cholesterol EpiScore. 233 CpGs described in
McCartney et al. (2018).
}
\details{
\code{get_coefMcCartneyLDL} retrieves this resource from ExperimentHub.
}
\examples{
get_coefMcCartneyLDL()
}
