% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefMcCartneySmoking}
\alias{get_coefMcCartneySmoking}
\title{coefMcCartneySmoking}
\usage{
get_coefMcCartneySmoking()
}
\value{
coefMcCartneySmoking dataset
}
\description{
Coefficients McCartney smoking EpiScore. 233 CpGs described in McCartney et
al. (2018).
}
\details{
\code{get_coefMcCartneySmoking} retrieves this resource from ExperimentHub.
}
\examples{
get_coefMcCartneySmoking()
}
