% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefMcCartneyTotalChol}
\alias{get_coefMcCartneyTotalChol}
\title{coefMcCartneyTotalChol}
\usage{
get_coefMcCartneyTotalChol()
}
\value{
coefMcCartneyTotalChol dataset
}
\description{
Coefficients McCartney total cholesterol EpiScore. 204 CpGs described in
McCartney et al. (2018).
}
\details{
\code{get_coefMcCartneyTotalChol} retrieves this resource from
ExperimentHub.
}
\examples{
get_coefMcCartneyTotalChol()
}
