% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefMcCigarette}
\alias{get_coefMcCigarette}
\title{coefMcCigarette}
\usage{
get_coefMcCigarette()
}
\value{
coefMcCigarette dataset
}
\description{
Coefficients mCigarette smoking score. 1255 CpGs described in Chybowska et
al. (2025).
}
\details{
\code{get_coefMcCigarette} retrieves this resource from ExperimentHub.
}
\examples{
get_coefMcCigarette()
}
