% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefNEOaPMA450K}
\alias{get_coefNEOaPMA450K}
\title{coefNEOaPMA450K}
\usage{
get_coefNEOaPMA450K()
}
\value{
coefNEOaPMA450K dataset
}
\description{
Coefficients NEOage post-menstrual age clock (450K). 408 CpGs described in
Graw et al. (2021).
}
\details{
\code{get_coefNEOaPMA450K} retrieves this resource from ExperimentHub.
}
\examples{
get_coefNEOaPMA450K()
}
