% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefNEOaPMAEPIC}
\alias{get_coefNEOaPMAEPIC}
\title{coefNEOaPMAEPIC}
\usage{
get_coefNEOaPMAEPIC()
}
\value{
coefNEOaPMAEPIC dataset
}
\description{
Coefficients NEOage post-menstrual age clock (EPIC). 520 CpGs described in
Graw et al. (2021).
}
\details{
\code{get_coefNEOaPMAEPIC} retrieves this resource from ExperimentHub.
}
\examples{
get_coefNEOaPMAEPIC()
}
