% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefNEOaPNA450K}
\alias{get_coefNEOaPNA450K}
\title{coefNEOaPNA450K}
\usage{
get_coefNEOaPNA450K()
}
\value{
coefNEOaPNA450K dataset
}
\description{
Coefficients NEOage post-natal age clock (450K). 302 CpGs described in Graw
et al. (2021).
}
\details{
\code{get_coefNEOaPNA450K} retrieves this resource from ExperimentHub.
}
\examples{
get_coefNEOaPNA450K()
}
