% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefNEOaPNAEPIC}
\alias{get_coefNEOaPNAEPIC}
\title{coefNEOaPNAEPIC}
\usage{
get_coefNEOaPNAEPIC()
}
\value{
coefNEOaPNAEPIC dataset
}
\description{
Coefficients NEOage post-natal age clock (EPIC). 508 CpGs described in Graw
et al. (2021).
}
\details{
\code{get_coefNEOaPNAEPIC} retrieves this resource from ExperimentHub.
}
\examples{
get_coefNEOaPNAEPIC()
}
