% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_coefPedBE}
\alias{get_coefPedBE}
\title{coefPedBE}
\usage{
get_coefPedBE()
}
\value{
coefPedBE dataset
}
\description{
\code{get_coefPedBE} returns the PedBE's clock coefficients
}
\examples{
get_coefPedBE()
}
