% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefRepliTali}
\alias{get_coefRepliTali}
\title{coefRepliTali}
\usage{
get_coefRepliTali()
}
\value{
coefRepliTali dataset
}
\description{
Coefficients RepliTali mitotic clock. 87 CpGs described in Endicott et al.
(2022).
}
\details{
\code{get_coefRepliTali} retrieves this resource from ExperimentHub.
}
\examples{
get_coefRepliTali()
}
