% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_coefSkin}
\alias{get_coefSkin}
\title{coefSkin}
\usage{
get_coefSkin()
}
\value{
coefSkin dataset
}
\description{
\code{get_coefSkin} returns the Horvath's skin+blood clock coefficients
}
\examples{
get_coefSkin()
}
