% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefStemTOC}
\alias{get_coefStemTOC}
\title{coefStemTOC}
\usage{
get_coefStemTOC()
}
\value{
coefStemTOC dataset
}
\description{
CpG set for the stemTOC mitotic counter. 371 CpGs described in Zhu et al.
(2024).
}
\details{
\code{get_coefStemTOC} retrieves this resource from ExperimentHub.
}
\examples{
get_coefStemTOC()
}
