% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefVidalBralo}
\alias{get_coefVidalBralo}
\title{coefVidalBralo}
\usage{
get_coefVidalBralo()
}
\value{
coefVidalBralo dataset
}
\description{
Coefficients Vidal-Bralo blood clock. 8 CpGs described in Vidal-Bralo et
al. (2016).
}
\details{
\code{get_coefVidalBralo} retrieves this resource from ExperimentHub.
}
\examples{
get_coefVidalBralo()
}
