% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData_2026.R
\name{get_coefWeidner}
\alias{get_coefWeidner}
\title{coefWeidner}
\usage{
get_coefWeidner()
}
\value{
coefWeidner dataset
}
\description{
Coefficients Weidner three-CpG clock. 3 CpGs described in Weidner et al.
(2014).
}
\details{
\code{get_coefWeidner} retrieves this resource from ExperimentHub.
}
\examples{
get_coefWeidner()
}
