% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_cpgs_bn}
\alias{get_cpgs_bn}
\title{cpgs_bn}
\usage{
get_cpgs_bn()
}
\value{
cpgs_bn dataset
}
\description{
\code{get_cpgs_bn} returns data to use Horvath's CpGs to train a 
Bayesian Neural Network (BNN)
}
\examples{
get_cpgs_bn()
}
