% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_MethylationDataExample}
\alias{get_MethylationDataExample}
\title{MethylationDataExample55}
\usage{
get_MethylationDataExample()
}
\value{
MethylationDataExample55 dataset
}
\description{
\code{get_MethylationDataExample} MethylationDataExample55 for vignette
}
\examples{
get_MethylationDataExample()
}
