% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_probeAnnotation21kdatMethUsed}
\alias{get_probeAnnotation21kdatMethUsed}
\title{probeAnnotation21kdatMethUsed}
\usage{
get_probeAnnotation21kdatMethUsed()
}
\value{
probeAnnotation21kdatMethUsed dataset
}
\description{
\code{get_probeAnnotation21kdatMethUsed} returns probeAnnotation21kdatMethUsed
}
\examples{
get_probeAnnotation21kdatMethUsed()
}
