% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_references}
\alias{get_references}
\title{references}
\format{
A list with different data frame.
  \describe{
    \item{\code{blood gse35069 complete}}{methylation profiles from
    Reinius 2012 for purified blood cell types.
    It includes CD4T, CD8T, Mono,}
    \item{\code{blood gse35069}}{methylation profiles from
    Reinius 2012 for purified blood cell types.
    It includes CD4T,CD8T, Mono, Bcell, NK and Gran.}
    \item{\code{blood gse35069 chen}}{methylation profiles
    from Chen 2017 blood cell types.
    It includes CD4T, CD8T, Mono, Bcell, NK, Neu and Eos.}
    \item{\code{andrews and bakulski cord blood}}{Cord blood
    reference from Bakulski 2016. It includes Bcell,
    CD4T, CD8T, Gran, Mono, NK and nRBC.}
    \item{\code{cord blood gse68456}}{Cord blood methylation profiles
    from De 2015. It includes CD4T, CD8T, Mono, Bcell, NK,
    Neu, Eos and RBC.}
    \item{\code{gervin and lyle cord blood}}{Cord blood reference generated
    by Kristina Gervin and Robert Lyle, available at `miffil` package.
    It includes CD14, Bcell, CD4T, CD8T, NK, Gran.}
    \item{\code{saliva gse48472}}{Reference generated from the multi-tissue
    pannel from Slieker 2013. It includes Buccal, CD4T, CD8T,
    Mono, Bcell, NK, Gran.}
    }
}
\usage{
get_references()
}
\value{
references dataset
}
\description{
The methylclockData package is a repository of a few public
datasets that needs the *methylclock* package to estimate chronological
and gestational DNA methylation (DNAm) age as well as biological age
using different methylation clocks.
}
\details{
\code{get_references} returns the file with references data
}
\examples{
get_references()
}
