% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{get_TestDataset}
\alias{get_TestDataset}
\title{TestDataset}
\usage{
get_TestDataset()
}
\value{
TestDataset dataset
}
\description{
\code{get_TestDataset} returns TestDataset
}
\examples{
get_TestDataset()
}
